Move parse_content_disposition to network.download
